fix: Fix WordPress termination of PHP in wp-settings.php
This commit introduces a Puppet manifest that fixes the termination of PHP in wp-settings.php file of a WordPress installation. The manifest executes a sed command to replace all instances of "phpp" with "php" in the specified file.

The path attribute in the exec resource specifies the directories where Puppet should look for the sed executable to execute the command.